fix: refactor exit survey hooks to use redux state and improve presentation logic
