+++ apps/main-app/ts/features/itwallet/common/hooks/useItwEidFeedbackBottomSheet.tsx
@@ -20,6 +20,11 @@ type ItwEidFeedbackBottomSheetProps = {
 
 /**
  * Hook to open the feedback bottom sheet for the EID reissuance feature.
+ *
+ * Unlike the other exit survey hooks, presentation is NOT auto-triggered
+ * internally: it's used both from a redux-driven trigger (WalletHomeScreen)
+ * and from a direct user action (ItwPresentationEidVerificationExpiredScreen),
+ * so `present` stays exposed for the caller to invoke explicitly.
  * @param onPrimaryAction - Optional primary action to be executed when the primary button is pressed.
  * @param onSecondaryAction - Optional secondary action to be executed when the secondary button is pressed.
  */
